Credentials to model dir, now mapping to Bson document for mongodb
