+++ auth-server/src/main/java/pl/edu/agh/dp/tkgk/oauth2server/database/RAMDBFacade.java
@@ -2,7 +2,7 @@ package pl.edu.agh.dp.tkgk.oauth2server.database;
 
 import com.auth0.jwt.interfaces.DecodedJWT;
 import pl.edu.agh.dp.tkgk.oauth2server.endpoints.authrequest.AuthorizationRequest;
-import pl.edu.agh.dp.tkgk.oauth2server.endpoints.authrequest.Credentials;
+import pl.edu.agh.dp.tkgk.oauth2server.model.Credentials;
 import pl.edu.agh.dp.tkgk.oauth2server.model.AuthCode;
 import pl.edu.agh.dp.tkgk.oauth2server.model.Client;
 import pl.edu.agh.dp.tkgk.oauth2server.model.Session;
