--- auth-server/src/main/java/pl/edu/agh/dp/tkgk/oauth2server/endpoints/authrequest/Credentials.java
@@ -1,23 +1,0 @@
-package pl.edu.agh.dp.tkgk.oauth2server.endpoints.authrequest;
-
-import java.util.Objects;
-
-public record Credentials(String login, String password) {
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-
-        Credentials that = (Credentials) o;
-
-        if (!Objects.equals(login, that.login)) return false;
-        return Objects.equals(password, that.password);
-    }
-
-    @Override
-    public int hashCode() {
-        int result = login != null ? login.hashCode() : 0;
-        result = 31 * result + (password != null ? password.hashCode() : 0);
-        return result;
-    }
-}
