+++ auth-server/src/main/java/pl/edu/agh/dp/tkgk/oauth2server/model/Credentials.java
@@ -1,0 +1,42 @@
+package pl.edu.agh.dp.tkgk.oauth2server.model;
+
+import org.bson.codecs.pojo.annotations.BsonCreator;
+import org.bson.codecs.pojo.annotations.BsonId;
+import org.bson.codecs.pojo.annotations.BsonProperty;
+
+import java.util.Objects;
+
+public record Credentials(@BsonId() String login, @BsonProperty("password") String password) {
+
+    @BsonCreator
+    public Credentials(@BsonProperty("login") String login, @BsonProperty("password") String password) {
+        this.login = login;
+        this.password = password;
+    }
+
+    public String getLogin() {
+        return login;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+
+        Credentials that = (Credentials) o;
+
+        if (!Objects.equals(login, that.login)) return false;
+        return Objects.equals(password, that.password);
+    }
+
+    @Override
+    public int hashCode() {
+        int result = login != null ? login.hashCode() : 0;
+        result = 31 * result + (password != null ? password.hashCode() : 0);
+        return result;
+    }
+}
