+++ auth-server/src/test/java/pl/edu/agh/dp/tkgk/oauth2server/endpoints/authrequest/AuthorizationRequestMother.java
@@ -1,7 +1,6 @@
 package pl.edu.agh.dp.tkgk.oauth2server.endpoints.authrequest;
 
-import pl.edu.agh.dp.tkgk.oauth2server.endpoints.authrequest.AuthorizationRequest;
-import pl.edu.agh.dp.tkgk.oauth2server.endpoints.authrequest.Credentials;
+import pl.edu.agh.dp.tkgk.oauth2server.model.Credentials;
 import pl.edu.agh.dp.tkgk.oauth2server.model.util.CodeChallengeMethod;
 
 import java.util.Set;
